Fixing some numpy depracations (can't use float as input to np.zeros)

+++ PulseInterface.py
@@ -18,10 +18,10 @@ def make_pulse(sampling_rate, global_onset, global_offset, params_list):
         if len(t) > len(longest_t):
             longest_t = t
 
-    pulse_matrix = np.zeros((len(pulses), len(longest_t) + (global_onset + global_offset) * sampling_rate))
+    pulse_matrix = np.zeros((len(pulses), len(longest_t) + int((global_onset + global_offset) * sampling_rate)))
 
     for p, pulse in enumerate(pulses):
-        pulse_matrix[p][(global_onset * sampling_rate):(global_onset * sampling_rate)+len(pulse)] = pulse
+        pulse_matrix[p][int(global_onset * sampling_rate):int(global_onset * sampling_rate)+len(pulse)] = pulse
 
     t = np.linspace(0, pulse_matrix.shape[1] / sampling_rate, pulse_matrix.shape[1])
